LayoutTests/platform/ios-simulator/ios/touch/construct-TouchList-expected.txt

Test constructor for TouchList. <rdar://problem/6481690>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL TouchList.prototype.toString.call(new TouchList) should be [object TouchList]. Was [object Touch].
FAIL TouchList.prototype.toString.call(new TouchList(touch1)) should be [object TouchList]. Was [object Touch].
FAIL TouchList.prototype.toString.call(new TouchList(touch1, touch2)) should be [object TouchList]. Was [object Touch].
FAIL TouchList.prototype.toString.call(new TouchList(touch1, touch2, touch3)) should be [object TouchList]. Was [object Touch].
PASS successfullyParsed is true

TEST COMPLETE
